+++ Sources/Sentry/SentyOptionsInternal.m
@@ -12,6 +12,7 @@
 #import "SentryOptions.h"
 #import "SentryOptionsInternal.h"
 #import "SentrySessionReplayIntegration.h"
+#import "SentryStdoutLogIntegration.h"
 #import "SentrySwift.h"
 #import "SentrySwiftAsyncIntegration.h"
 
@@ -54,7 +55,7 @@
         [SentryANRTrackingIntegration class], [SentryAutoBreadcrumbTrackingIntegration class],
         [SentryAutoSessionTrackingIntegration class], [SentryCoreDataTrackingIntegration class],
         [SentryFileIOTrackingIntegration class], [SentryNetworkTrackingIntegration class],
-        [SentrySwiftAsyncIntegration class], nil];
+        [SentryStdOutLogIntegration class], [SentrySwiftAsyncIntegration class], nil];
 
 #if TARGET_OS_IOS && SENTRY_HAS_UIKIT
     [defaultIntegrations addObject:[SentryUserFeedbackIntegration class]];