+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -168,6 +168,7 @@
 #import "SentrySpotlightTransport.h"
 #import "SentryStacktrace.h"
 #import "SentryStacktraceBuilder.h"
+#import "SentryStdOutLogIntegration.h"
 #import "SentrySubClassFinder.h"
 #import "SentrySwift.h"
 #import "SentrySwiftAsyncIntegration.h"